Move DonationViewModal to common/components and guard Badge against null status
DonationViewModal is used across three pages (donations, donors, dashboard)
so it belongs in common/components/organisms per project conventions.
Also fixes Badge crash when receipt_status is null/undefined.
