+++ src/common/components/organisms/DonationViewModal.jsx
@@ -6,11 +6,11 @@ import { RECEIPT_SUBJECT, buildReceiptMessage } from '@/utils/receiptTemplate';
 import { X } from 'lucide-react';
 import PropTypes from 'prop-types';
 
-import DonationDeleteSection from './DonationDeleteSection';
-import DonationDonorInfoSection from './DonationDonorInfoSection';
-import DonationModalFooter from './DonationModalFooter';
-import DonationPaymentSection from './DonationPaymentSection';
-import DonationReceiptSection from './DonationReceiptSection';
+import DonationDeleteSection from '@/pages/donations/DonationDeleteSection';
+import DonationDonorInfoSection from '@/pages/donations/DonationDonorInfoSection';
+import DonationModalFooter from '@/pages/donations/DonationModalFooter';
+import DonationPaymentSection from '@/pages/donations/DonationPaymentSection';
+import DonationReceiptSection from '@/pages/donations/DonationReceiptSection';
 import {
   closeBtn,
   errorStyle,
@@ -18,9 +18,9 @@ import {
   modal,
   overlay,
   titleStyle,
-} from './DonationViewModal.styles';
-import EmailPreviewModal from './EmailPreviewModal';
-import { EMPTY_DONATION_VIEW_FORM, fromDonationRow } from './donationViewForm';
+} from '@/pages/donations/DonationViewModal.styles';
+import EmailPreviewModal from '@/pages/donations/EmailPreviewModal';
+import { EMPTY_DONATION_VIEW_FORM, fromDonationRow } from '@/pages/donations/donationViewForm';
 
 export default function DonationViewModal({
   open,
